@@ -18,6 +18,7 @@
   "dependencies": {
     "@egjs/hammerjs": "^2.0.17",
     "@lol768/jquery-querybuilder-no-eval": "^2.6.0",
+    "@testing-library/dom": "^10.0.0",
     "bootstrap": "4.6",
     "bootstrap-datepicker": "^1.9.0",
     "bootstrap-html5sortable": "^1.0.0",
@@ -37,11 +38,11 @@
     "moment": "^2.30.1",
     "popper.js": "^1.16.1",
     "propagating-hammerjs": "^2.0.0",
-    "react": "^16.13.1",
-    "react-app-polyfill": "^1.0.6",
-    "react-dom": "^16.13.1",
-    "react-grid-layout": "^0.18.3",
-    "react-modal": "^3.11.2",
+    "react": "^19.1.0",
+    "react-app-polyfill": "^3.0.0",
+    "react-dom": "^19.1.0",
+    "react-grid-layout": "^1.5.1",
+    "react-modal": "^3.16.3",
     "regenerator-runtime": "^0.14.1",
     "summernote": "^0.9.1",
     "tippy.js": "^6.3.7",
@@ -60,15 +61,15 @@
     "@babel/preset-typescript": "^7.27.0",
     "@babel/runtime-corejs3": "^7.27.0",
     "@jest/globals": "^29.7.0",
-    "@testing-library/react": "12",
+    "@testing-library/react": "^16.3.0",
     "@types/bootstrap-datepicker": "^1.10.0",
     "@types/form-serialize": "^0.7.4",
     "@types/jest": "^29.5.14",
     "@types/jquery": "^3.5.32",
     "@types/jstree": "^3.3.46",
-    "@types/react": "^17.0.41",
-    "@types/react-dom": "^17.0.14",
-    "@types/react-grid-layout": "^1.3.2",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
+    "@types/react-grid-layout": "^1.3.5",
     "@types/react-modal": "^3.16.3",
     "@types/typeahead.js": "^0.11.6",
     "@webpack-cli/serve": "^3.0.1",
