@@ -1,32 +1,19 @@
 import { Component } from 'component'
-import "react-app-polyfill/stable";
-
-import "core-js/es/array/is-array";
-import "core-js/es/map";
-import "core-js/es/set";
-import "core-js/es/object/define-property";
-import "core-js/es/object/keys";
-import "core-js/es/object/set-prototype-of";
-
-import "./react/polyfills/classlist";
 
 import React from "react";
-import ReactDOM from "react-dom";
-import App from "./react/app";
+import ReactDOM from "react-dom/client";
+import App from "./react/App";
 import ApiClient from "./react/api";
+import { ReactGridLayoutProps } from 'react-grid-layout';
 
-/**
- * Dashboard component which can contain multiple widgets.
- */
 class DashboardComponent extends Component {
-  /**
-   * Create a new Dashboard
-   * @param {HTMLElement} element The element to attach the dashboard to.
-   */
-  constructor(element) {
+  el: JQuery<HTMLElement>;
+  gridConfig: ReactGridLayoutProps;
+
+  constructor(element:HTMLElement)  {
     super(element)
     this.el = $(this.element)
-
+    
     this.gridConfig = {
       cols: 2,
       margin: [32, 32],
@@ -37,19 +24,18 @@ class DashboardComponent extends Component {
     this.initDashboard()
   }
 
-  /**
-   * Initialize the dashboard.
-   */
   initDashboard() {
     this.element.className = "";
     const widgetsEls = Array.prototype.slice.call(document.querySelectorAll("#ld-app > div"));
-    const widgets = widgetsEls.map(el => ({
+    const widgets = widgetsEls.map((el: HTMLElement) => ({
       html: el.innerHTML,
       config: JSON.parse(el.getAttribute("data-grid")),
     }));
     const api = new ApiClient(this.element.getAttribute("data-dashboard-endpoint") || "");
 
-    ReactDOM.render(
+    const root = ReactDOM.createRoot(this.element);
+
+    root.render(
       <App
         widgets={widgets}
         dashboardId={this.element.getAttribute("data-dashboard-id")}
@@ -60,9 +46,8 @@ class DashboardComponent extends Component {
         noDownload={this.element.getAttribute("data-dashboard-no-download") === "true"}
         api={api}
         widgetTypes={JSON.parse(this.element.getAttribute("data-widget-types") || "[]")}
-        dashboards={JSON.parse(this.element.getAttribute("data-dashboards") || "[]")}
-        gridConfig={this.gridConfig} />,
-      this.element,
+        dashboards={JSON.parse(this.element.getAttribute("data-dashboards") || "[]" )}
+        gridConfig={this.gridConfig} />
     );
   }
 }
