@@ -1,39 +1,33 @@
 import React from "react";
+import { FooterProps } from "./types";
+import { Dropdown } from "react-bootstrap";
 
-/**
- * Create a footer component for the dashboard
- * @param param0 The props for the footer component
- */
-const Footer = ({ addWidget, widgetTypes, currentDashboard, readOnly, noDownload }: { addWidget: (type: string) => void, widgetTypes: string[], currentDashboard: any, readOnly: boolean, noDownload: boolean }) => {
+const Footer = ({ addWidget, widgetTypes, currentDashboard, readOnly, noDownload }: FooterProps) => {
   return (
     <div className='ld-footer-container'>
-      {noDownload ? null : <div className="btn-group mb-3 mb-md-0 mr-md-4">
-        <button type="button" className="btn btn-default dropdown-toggle" data-toggle="dropdown" aria-expanded="false" aria-controls="menu_view">
-          Download <span className="caret"></span>
-        </button>
-        <div className="dropdown-menu dropdown__menu dropdown-menu-right scrollable-menu" role="menu">
-          <ul id="menu_view" className="dropdown__list">
-            <li className="dropdown__item">
-              <a className="link link--plain" href={currentDashboard.download_url}>As PDF</a>
-            </li>
-          </ul>
-        </div>
-      </div>}
+      {noDownload ||
+        <Dropdown>
+          <Dropdown.Toggle variant="default" id="dropdown-download">
+            Download
+          </Dropdown.Toggle>
 
-      {readOnly ? null : <div className="btn-group">
-        <button type="button" className="btn btn-default dropdown-toggle" data-toggle="dropdown" aria-expanded="false" aria-controls="menu_view">
-          Add Widget
-        </button>
-        <div className="dropdown-menu dropdown__menu dropdown-menu-right scrollable-menu" role="menu">
-          <ul id="menu_view" className="dropdown__list">
+          <Dropdown.Menu>
+            <Dropdown.Item href={currentDashboard.download_url}>As PDF</Dropdown.Item>
+          </Dropdown.Menu>
+        </Dropdown>}
+      
+      {readOnly || 
+        <Dropdown>
+          <Dropdown.Toggle variant="default" id="dropdown-add-widget">
+            Add Widget
+          </Dropdown.Toggle>
+
+          <Dropdown.Menu>
             {widgetTypes.map(type => (
-              <li key={type} className="dropdown__item">
-                <a className="link link--plain" href="#" onClick={(e) => { e.preventDefault(); addWidget(type) }}>{type}</a>
-              </li>
+              <Dropdown.Item key={type} onClick={(e) => {e.preventDefault(); addWidget(type)}}>{type}</Dropdown.Item>
             ))}
-          </ul>
-        </div>
-      </div>}
+          </Dropdown.Menu>
+        </Dropdown>}
     </div>
   );
 };
